% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_annotation.R
\name{append_annotation}
\alias{append_annotation}
\title{Add annotation from extra annotation file}
\usage{
append_annotation(input, d)
}
\arguments{
\item{d}{a DE result data.frame including either ensembl_gene_id as a column.}
}
\value{
data.frame
}
\description{
Add annotation from extra annotation file
}
