% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_secondary_data.R
\name{append_secondary_data}
\alias{append_secondary_data}
\title{Appending secondary DE results to primary data set.}
\usage{
append_secondary_data(
  input,
  data,
  de_results = gene_results_de(),
  ensembl_lift = "human_mouse_ensembl.txt"
)
}
\arguments{
\item{data}{primary data set on which to append secondary DE results}

\item{de_results}{a reactive data frame including the secondary DE results}

\item{ensembl_lift}{a text file to translate ensembl ids between species}
}
\value{
data.frame
}
\description{
Appending secondary DE results to primary data set.
}
