% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count2deseq_analysis.R
\name{count2deseq_analysis}
\alias{count2deseq_analysis}
\title{DEseq2 analysis of raw count data}
\usage{
count2deseq_analysis(input, countdata, pheno)
}
\arguments{
\item{countdata}{an N x M data.frame with N genes and M samples}

\item{pheno}{an A x B phenotype data.frame with A samples, column 2 will include batch if present.}
}
\value{
named list, including the variables "test", "norm_counts", "dds", "phenotypes"
}
\description{
Includes option of batch correction
}
