% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_circ.R
\name{de_circ}
\alias{de_circ}
\title{Differential Expression analysis on circRNA}
\usage{
de_circ(data, pheno)
}
\arguments{
\item{data}{count data}

\item{pheno}{phenotype dataframe}
}
\value{
list of results
}
\description{
Structures the data, and calls the count2deseq_analysis function
}
