% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_results.R
\name{filter_results}
\alias{filter_results}
\title{Filter DE results on log2foldchange and p-value, and includes 10 percent of "non-significant"}
\usage{
filter_results(input, data)
}
\arguments{
\item{data}{the dateframe}
}
\value{
the filtered dataframe
}
\description{
Filter DE results on log2foldchange and p-value, and includes 10 percent of "non-significant"
}
