% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_d.R
\name{input_d}
\alias{input_d}
\title{Reads in the count files
Handles if the files are separate for each sample, the zipped folder is unzipped and the data combined.}
\usage{
input_d(input)
}
\arguments{
\item{input}{UI data including the amount datasets (nfiles)}
}
\value{
list of "count", "pheno" and "circ" (if circRNA data is included)
}
\description{
Reads in the count files
Handles if the files are separate for each sample, the zipped folder is unzipped and the data combined.
}
