% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_analysis.R
\name{limma_analysis}
\alias{limma_analysis}
\title{Sets up a fitting regression model and runs a limma analysis}
\usage{
limma_analysis(countdata_norm, phenotypes)
}
\arguments{
\item{countdata_norm}{normalized count data}

\item{phenotypes}{a vector of phenotype ids}
}
\value{
list of results
}
\description{
Sets up a fitting regression model and runs a limma analysis
}
