% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\title{function used by input_d
if files are not combined this reads from a folder containing the single samples}
\usage{
read_files(input, dir)
}
\arguments{
\item{dir}{the directory of the folder}

\item{d}{the delimiter}
}
\value{
the combined dataset
}
\description{
function used by input_d
if files are not combined this reads from a folder containing the single samples
}
