% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pheno.R
\name{read_pheno}
\alias{read_pheno}
\title{Read the file including phenotypes}
\usage{
read_pheno(pheno)
}
\arguments{
\item{pheno}{the path to the pheontype file}
}
\value{
a dataframe with the phenotypes
}
\description{
Read the file including phenotypes
}
