% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_Atlas.R
\name{search_Atlas}
\alias{search_Atlas}
\title{Search the Atlas database}
\usage{
search_Atlas(input, atlas_id = F, probe_library = probe_library())
}
\arguments{
\item{input}{from UI, inclludes species, atlas id, atlas query}

\item{probe_library}{a library translating different ids}
}
\value{
data.frame of DE analysis
}
\description{
Search the Atlas database
}
