% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGenes.R
\name{testGenes}
\alias{testGenes}
\title{Connects to the Atlas API, and tests for overlapping expression}
\usage{
testGenes(gene_vector, ebi_id, input)
}
\arguments{
\item{A}{vector of ensembl gene ids to test against the database}
}
\value{
data.frame including dataset id and link
}
\description{
Connects to the Atlas API, and tests for overlapping expression
}
