% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_atlas.R
\name{update_atlas}
\alias{update_atlas}
\title{Append data from the Atlas db to dataset}
\usage{
update_atlas(atlas_id = F, gene_data, input)
}
\arguments{
\item{gene_data}{the dataset}

\item{input}{}
}
\description{
Append data from the Atlas db to dataset
}
